# Backend/Synonyms_Test/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Synonyms_Test, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Synonym/GetSynonyms": {
      "post": {
        "tags": [
          "Synonym"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GetSynonymsDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetSynonymsResponseDto"
              }
            }
          }
        }
      }
    },
    "/api/Synonym/AddNewWord": {
      "post": {
        "tags": [
          "Synonym"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddSynonymRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Synonym/GetAllWords": {
      "get": {
        "tags": [
          "Synonym"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetAllWordsResponseDto"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AddSynonymRequestDto": {
      "type": "object",
      "properties": {
        "newWord": {
          "type": "string"
        },
        "synonym": {
          "type": "string"
        }
      }
    },
    "GetAllWordsResponseDto": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "GetSynonymsDto": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "GetSynonymsResponseDto": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    }
  }
}